requirements.txt
dotenv>=0.9.9
python-dotenv>=1.1.0
google-genai==1.11.0
gradio
google-cloud-secret-manager